+++ packages/components/src/AutoComplete/AutoComplete.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import omit from 'rc-util/lib/omit';
 import toArray from '../utils/toArray';
 
-import { Select, InternalSelectProps, OptionType, RefSelectProps } from '../Select/Select';
+import { Select, OptionType, RefSelectProps } from '../Select/Select';
+import { InternalSelectProps } from '../Select/Select.styles';
 
 const { Option } = Select;