+++ packages/components/src/Select/Select.tsx
@@ -1,7 +1,6 @@
 // https://github.dev/ant-design/ant-design/blob/756dc8f130547277a28548c5b46658ff46f131f9/components/select/index.tsx
 import React from 'react';
-import { Option as SelectOption, OptGroup as SelectOptGroup } from 'rc-select';
-import { OptionProps } from 'rc-select/lib/Option';
+import { Option, OptGroup } from 'rc-select';
 import forwardRef from '../utils/forwardRef';
 import getIcons from './iconUtil';
 import {
@@ -14,10 +13,7 @@ import {
 
 type RawValue = string | number;
 
-// @ts-ignore
-export { OptionProps, InternalSelectProps };
-
-export type OptionType = typeof SelectOption;
+export type OptionType = typeof Option;
 
 export interface LabeledValue {
   key?: string;
@@ -97,7 +93,7 @@ const InternalSelect = forwardRef<SelectProps<any>, 'div'>(
           notFoundContent={mergedNotFound}
           getPopupContainer={getPopupContainer}
           prefixCls="kubed-select"
-          bordered={bordered}
+          $bordered={bordered}
         />
       </>
     );
@@ -108,16 +104,14 @@ type InternalSelectType = typeof InternalSelect;
 
 interface SelectInterface extends InternalSelectType {
   SECRET_COMBOBOX_MODE_DO_NOT_USE: string;
-  Option: typeof SelectOption;
-  OptGroup: typeof SelectOptGroup;
+  Option: typeof Option;
+  OptGroup: typeof OptGroup;
 }
 
 export const Select = InternalSelect as SelectInterface;
 
 Select.SECRET_COMBOBOX_MODE_DO_NOT_USE = SECRET_COMBOBOX_MODE_DO_NOT_USE;
-
-Select.Option = SelectOption;
-
-Select.OptGroup = SelectOptGroup;
+Select.Option = Option;
+Select.OptGroup = OptGroup;
 
 Select.displayName = '@kubed/components/Select';