+++ packages/components/src/TimePicker/locales/zh-tw.ts
@@ -1,0 +1,7 @@
+import { TimePickerLocale } from '../TimePicker';
+
+const locale: TimePickerLocale = {
+  placeholder: '請選擇時間',
+};
+
+export default locale;