+++ docs/CHANGELOG.md
@@ -1,5 +1,134 @@
 # kubed-documents
 
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [[`f5ed279`](https://github.com/kubesphere/kube-design/commit/f5ed279be5140890fd91e762c02593988576ca46)]:
+  - @kubed/hooks@0.0.7
+  - @kubed/components@0.0.26
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [[`db858ee`](https://github.com/kubesphere/kube-design/commit/db858ee70bc841176d9dd8cf3d52ce6a3927fc2c)]:
+  - @kubed/hooks@0.0.6
+  - @kubed/components@0.0.25
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [[`f39c7e8`](https://github.com/kubesphere/kube-design/commit/f39c7e8d9550ddd329dc057bf6a737835b738d33)]:
+  - @kubed/components@0.0.24
+  - @kubed/hooks@0.0.5
+
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [[`5358eca`](https://github.com/kubesphere/kube-design/commit/5358eca08e547b35870d15258e3644e4bee0ed92)]:
+  - @kubed/components@0.0.23
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [[`bf37ccc`](https://github.com/kubesphere/kube-design/commit/bf37cccbf64598b67b6e599b61c9d4ed33ec8504)]:
+  - @kubed/components@0.0.22
+  - @kubed/hooks@0.0.4
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [[`7c7e066`](https://github.com/kubesphere/kube-design/commit/7c7e066d5a612344865f8279d39582e65c71919a)]:
+  - @kubed/components@0.0.21
+
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [[`e3b1a57`](https://github.com/kubesphere/kube-design/commit/e3b1a57e6adb6480963399e30151978511d89350)]:
+  - @kubed/components@0.0.20
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [[`dc41c69`](https://github.com/kubesphere/kube-design/commit/dc41c69b266e08345d7c43222bee5d1fabe10087)]:
+  - @kubed/components@0.0.19
+
+## 0.0.18
+
+### Patch Changes
+
+- [`a7f7e6a`](https://github.com/kubesphere/kube-design/commit/a7f7e6aa089002b1c24ab650aef124a7ecc3157d) Thanks [@chenz24](https://github.com/chenz24)! - 1. Use classnames instead of clsx.
+  2. Fix package bug cause by importing dayjs locales
+- Updated dependencies [[`a7f7e6a`](https://github.com/kubesphere/kube-design/commit/a7f7e6aa089002b1c24ab650aef124a7ecc3157d)]:
+  - @kubed/components@0.0.18
+
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [[`6fc0e96`](https://github.com/kubesphere/kube-design/commit/6fc0e96fb62e75eb9f1eb5c79347896f8c971cc1), [`eb0b852`](https://github.com/kubesphere/kube-design/commit/eb0b8520f105ebd91d25a0b9a7b8ab3c6a86b7fc)]:
+  - @kubed/components@0.0.17
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [[`0f4d178`](https://github.com/kubesphere/kube-design/commit/0f4d17860c1ae0c5ebcfdcd61b3b2288df7fb877)]:
+  - @kubed/components@0.0.16
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [[`c619355`](https://github.com/kubesphere/kube-design/commit/c6193554484524736c95bf263a1af1a6fdd394e8)]:
+  - @kubed/components@0.0.15
+
+## 0.0.14
+
+### Patch Changes
+
+- [`81610f7`](https://github.com/kubesphere/kube-design/commit/81610f7b72e85c675dfd129bd5c676ad09c6f702) Thanks [@chenz24](https://github.com/chenz24)! - 1. Fix ModalContext default value.
+  2. Fix MenuItem error cause by 'as' props.
+  3. Fix background of doc page in dark mode.
+- Updated dependencies [[`81610f7`](https://github.com/kubesphere/kube-design/commit/81610f7b72e85c675dfd129bd5c676ad09c6f702)]:
+  - @kubed/components@0.0.14
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`e1fe306`](https://github.com/kubesphere/kube-design/commit/e1fe30612f48839ae6eb40522ba28dd893ccbf9f)]:
+  - @kubed/components@0.0.13
+  - @kubed/icons@0.0.2
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`0cff4d4`](https://github.com/kubesphere/kube-design/commit/0cff4d4ea509b874b0a746b8a4dbaa24f7955c62)]:
+  - @kubed/components@0.0.12
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`c8be7d3`](https://github.com/kubesphere/kube-design/commit/c8be7d31582ee31fa7c33ee89df5d36dd659ab36)]:
+  - @kubed/components@0.0.11
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`c2610f8`](https://github.com/kubesphere/kube-design/commit/c2610f898a081e743918bc755a308ae2c2016035)]:
+  - @kubed/components@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes