+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @kubed/hooks
 
+## 0.0.7
+
+### Patch Changes
+
+- [`f5ed279`](https://github.com/kubesphere/kube-design/commit/f5ed279be5140890fd91e762c02593988576ca46) Thanks [@chenz24](https://github.com/chenz24)! - 1. Add useUnmount hook.
+
+## 0.0.6
+
+### Patch Changes
+
+- [`db858ee`](https://github.com/kubesphere/kube-design/commit/db858ee70bc841176d9dd8cf3d52ce6a3927fc2c) Thanks [@chenz24](https://github.com/chenz24)! - 1. add useDidUpdate hook.
+
+## 0.0.5
+
+### Patch Changes
+
+- [`f39c7e8`](https://github.com/kubesphere/kube-design/commit/f39c7e8d9550ddd329dc057bf6a737835b738d33) Thanks [@chenz24](https://github.com/chenz24)! - 1. Completely Refactor Empty component.
+  2. Add useDebouncedValue hook.
+  3. Reset default props for LoadingOverlay.
+
+## 0.0.4
+
+### Patch Changes
+
+- [`bf37ccc`](https://github.com/kubesphere/kube-design/commit/bf37cccbf64598b67b6e599b61c9d4ed33ec8504) Thanks [@chenz24](https://github.com/chenz24)! - 1. Add hotkey hook.
+  2. Fix Banner cannot set className
+
 ## 0.0.3
 
 ### Patch Changes